requirements.txt
boto3
pymongo
tensorboard
ruamel.yaml